mass_density moved to densities.py
