+++ plasmapy/formulary/tests/test_misc.py
@@ -11,10 +11,8 @@ from plasmapy.formulary.misc import (
     DB_,
     magnetic_energy_density,
     magnetic_pressure,
-    mass_density,
     pmag_,
     pth_,
-    rho_,
     thermal_pressure,
     ub_,
 )
@@ -36,66 +34,12 @@ T_e = 1e6 * u.K
         (DB_, Bohm_diffusion),
         (ub_, magnetic_energy_density),
         (pmag_, magnetic_pressure),
-        (rho_, mass_density),
         (pth_, thermal_pressure),
     ],
 )
 def test_aliases(alias, parent):
     """Test all aliases defined in misc.py"""
     assert alias is parent
-
-
-class Test_mass_density:
-    r"""Test the mass_density function in misc.py."""
-
-    @pytest.mark.parametrize(
-        ("args", "kwargs", "conditional"),
-        [
-            ((-1 * u.kg * u.m**-3, "He"), {}, pytest.raises(ValueError)),
-            ((-1 * u.m**-3, "He"), {}, pytest.raises(ValueError)),
-            (("not a Quantity", "He"), {}, pytest.raises(TypeError)),
-            ((1 * u.m**-3,), {}, pytest.raises(TypeError)),
-            ((1 * u.J, "He"), {}, pytest.raises(u.UnitTypeError)),
-            ((1 * u.m**-3, None), {}, pytest.raises(TypeError)),
-            (
-                (1 * u.m**-3, "He"),
-                {"z_ratio": "not a ratio"},
-                pytest.raises(TypeError),
-            ),
-        ],
-    )
-    def test_raises(self, args, kwargs, conditional):
-        with conditional:
-            mass_density(*args, **kwargs)
-
-    @pytest.mark.parametrize(
-        ("args", "kwargs", "expected"),
-        [
-            ((1.0 * u.g * u.m**-3, ""), {}, 1.0e-3 * u.kg * u.m**-3),
-            ((5.0e12 * u.cm**-3, "He"), {}, 3.32323849e-8 * u.kg * u.m**-3),
-            (
-                (5.0e12 * u.cm**-3, Particle("He")),
-                {},
-                3.32323849e-8 * u.kg * u.m**-3,
-            ),
-            (
-                (5.0e12 * u.cm**-3, "He"),
-                {"z_ratio": 0.5},
-                1.66161925e-08 * u.kg * u.m**-3,
-            ),
-            (
-                (5.0e12 * u.cm**-3, "He"),
-                {"z_ratio": -0.5},
-                1.66161925e-08 * u.kg * u.m**-3,
-            ),
-        ],
-    )
-    def test_values(self, args, kwargs, expected):
-        assert np.isclose(mass_density(*args, **kwargs), expected)
-
-    def test_handle_nparrays(self):
-        """Test for ability to handle numpy array quantities"""
-        assert_can_handle_nparray(mass_density)
 
 
 def test_thermal_pressure():
